requirements.txt
# ─────────────────────────────
# CORE AI / LLM
# ─────────────────────────────
anthropic>=0.34.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ─────────────────────────────
# SEARCH / RESEARCH
# ─────────────────────────────
tavily-python>=0.3.0
exa-py>=1.0.0

# ─────────────────────────────
# DOCUMENT GENERATION
# ─────────────────────────────
python-docx>=1.1.0

# ─────────────────────────────
# WEB BACKEND (CRITICAL)
# ─────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# ─────────────────────────────
# ENV / CONFIG
# ─────────────────────────────
python-dotenv>=1.0.0

# ─────────────────────────────
# HTTP / SCRAPING
# ─────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0

# ─────────────────────────────
# AUTO APPLY (browser automation)
# ─────────────────────────────
playwright>=1.44.0

# ─────────────────────────────
# EMAIL MONITOR (real-time OTP)
# ─────────────────────────────
# imaplib is stdlib — no extra install needed

# ─────────────────────────────
# DATA
# ─────────────────────────────
pandas>=2.2.0